Further progress on the new installer.  We are only a few short days away from being able to try it out.
TODO:  Make it so most checks are only run at the top of steps 2, 7, and 8.

TODO:  Add a check to the top of step 4 that looks for an existing installation in the database and raises any necessary warnings.

TODO:  Add a field generator to collect user input in steps 3, 4, 5, and 6. 

MAYBE-TODO:  Back off of the Perl Regular expressions and find an alternate field comparison, but maybe not..